+++ ESFramework/ESFramework.h
@@ -44,3 +44,4 @@
 #import "UIToolbar+ESAdditions.h"
 #import "UIView+ESAdditions.h"
 #import "UIViewController+ESAdditions.h"
+#import "UIWindow+ESAdditions.h"